+++ Assets/Scripts/Game/Services/PersonalityLogService.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 50a5f441ba36442d1ad4f5339d2d0591
